nunet/requirements.txt
docker
grpcio
grpcio-tools
protobuf
opencv-python
Pillow
SQLAlchemy
bcrypt
ipython
urllib3
psutil
pytest
pytest-cov
codecov
coverage
psycopg2 
selenium
pyvirtualdisplay

